Split handlers.rs into domain modules; surface swallowed errors
