+++ super-stt-daemon/src/daemon/mod.rs
@@ -1,14 +1,17 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
+pub mod backend_config_handlers;
 pub mod core;
 pub mod device_management;
+pub mod download_handlers;
 pub mod events;
-pub mod handlers;
 pub mod http;
 pub(crate) mod language;
 pub mod language_handlers;
 pub mod model_management;
 pub mod recording;
+pub mod settings_handlers;
+pub mod status_handlers;
 pub mod theme_handlers;
 pub mod transcription;
 pub mod types;
